+++ all_receitas/lib/var_json.dart
@@ -1,8 +1,8 @@
 final List<Map<String, dynamic>> menuItens = [
-  // {
-  //   'title': 'Receita 1 (1 a 9)',
-  //   'rota': '/re1_1'
-  // },
+  {
+    'title': 'Receita 1 (1 a 9)',
+    'rota': '/re1_1'
+  },
   {
     'title': 'Receita 1 (10 e 11)',
     'rota': '/re1_2'
